+++ sdmx-dl-desktop/src/main/java/internal/sdmxdl/desktop/DesktopWebFactory.java
@@ -1,15 +1,8 @@
 package internal.sdmxdl.desktop;
 
-import lombok.NonNull;
 import nbbrd.io.curl.CurlHttpURLConnection;
-import sdmxdl.web.Network;
+import sdmxdl.provider.web.SingleNetworkingSupport;
 import sdmxdl.web.SdmxWebManager;
-import sdmxdl.web.URLConnectionFactory;
-
-import javax.net.ssl.HostnameVerifier;
-import javax.net.ssl.HttpsURLConnection;
-import javax.net.ssl.SSLSocketFactory;
-import java.net.ProxySelector;
 
 public class DesktopWebFactory {
 
@@ -20,32 +13,8 @@ public class DesktopWebFactory {
     public static SdmxWebManager loadManager() {
         return SdmxWebManager.ofServiceLoader()
                 .toBuilder()
-                .network(getNetwork())
+                .networking(SingleNetworkingSupport.builder().id("CURL").urlConnectionFactoryOf(CurlHttpURLConnection::of).build())
                 .eventListener((source, msg) -> System.out.println(source.getId() + ": " + msg))
                 .build();
-    }
-
-    private static Network getNetwork() {
-        return new Network() {
-            @Override
-            public @NonNull ProxySelector getProxySelector() {
-                return ProxySelector.getDefault();
-            }
-
-            @Override
-            public @NonNull SSLSocketFactory getSSLSocketFactory() {
-                return HttpsURLConnection.getDefaultSSLSocketFactory();
-            }
-
-            @Override
-            public @NonNull HostnameVerifier getHostnameVerifier() {
-                return HttpsURLConnection.getDefaultHostnameVerifier();
-            }
-
-            @Override
-            public @NonNull URLConnectionFactory getURLConnectionFactory() {
-                return CurlHttpURLConnection::of;
-            }
-        };
     }
 }