+++ sdmx-dl-grpc/src/main/java/sdmxdl/grpc/GrpcWebFactory.java
@@ -1,15 +1,8 @@
 package sdmxdl.grpc;
 
-import lombok.NonNull;
 import nbbrd.io.curl.CurlHttpURLConnection;
-import sdmxdl.web.Network;
+import sdmxdl.provider.web.SingleNetworkingSupport;
 import sdmxdl.web.SdmxWebManager;
-import sdmxdl.web.URLConnectionFactory;
-
-import javax.net.ssl.HostnameVerifier;
-import javax.net.ssl.HttpsURLConnection;
-import javax.net.ssl.SSLSocketFactory;
-import java.net.ProxySelector;
 
 @lombok.experimental.UtilityClass
 class GrpcWebFactory {
@@ -17,31 +10,7 @@ class GrpcWebFactory {
     public static SdmxWebManager loadManager() {
         return SdmxWebManager.ofServiceLoader()
                 .toBuilder()
-                .network(getNetwork())
+                .networking(SingleNetworkingSupport.builder().id("CURL").urlConnectionFactoryOf(CurlHttpURLConnection::of).build())
                 .build();
-    }
-
-    private static Network getNetwork() {
-        return new Network() {
-            @Override
-            public @NonNull ProxySelector getProxySelector() {
-                return ProxySelector.getDefault();
-            }
-
-            @Override
-            public @NonNull SSLSocketFactory getSSLSocketFactory() {
-                return HttpsURLConnection.getDefaultSSLSocketFactory();
-            }
-
-            @Override
-            public @NonNull HostnameVerifier getHostnameVerifier() {
-                return HttpsURLConnection.getDefaultHostnameVerifier();
-            }
-
-            @Override
-            public @NonNull URLConnectionFactory getURLConnectionFactory() {
-                return CurlHttpURLConnection::of;
-            }
-        };
     }
 }